Remove api
