--- apps/skills-microservice/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "skills-microservice",
-  "version": "1.0.0",
-  "description": "A microservice to keep track of your current skills",
-  "main": "serverless.ts",
-  "scripts": {
-    "build": "tsc",
-    "deploy": "serverless deploy --verbose",
-    "dev": "sls offline",
-    "test": "jest -c jest.config.ts --runInBand"
-  },
-  "dependencies": {
-    "aws-sdk": "^2.1058.0"
-  },
-  "devDependencies": {
-    "@serverless/typescript": "^2.16.0",
-    "@types/aws-lambda": "^8.10.64",
-    "@types/node": "^14.14.6",
-    "aws-sdk-mock": "^5.6.2",
-    "jest": "^27.5.0",
-    "jest-config": "^27.5.0",
-    "serverless": "^3.14.0",
-    "serverless-dynamodb-local": "^0.2.40",
-    "serverless-offline": "^8.3.1",
-    "serverless-plugin-typescript": "^2.1.0",
-    "ts-jest": "^27.1.3",
-    "ts-loader": "^8.0.10",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.5.4"
-  },
-  "author": "",
-  "license": "MIT"
-}
