--- apps/skills-microservice/src/functions/index.ts
@@ -1,4 +1,0 @@
-export * from './createSkills/handler'
-export * from './deleteSkills/handler'
-export * from './getSkills/handler'
-export * from './updateSkills/handler'
