--- apps/skills-microservice/swagger.yml
@@ -1,204 +1,0 @@
-openapi: 3.1.0
-info:
-  title: Skills Microservice
-  version: '1.0'
-  summary: A microservice to keep track of your current skills
-servers:
-  - url: 'http://localhost:8080'
-paths:
-  '/skills/{userId}':
-    post:
-      summary: Create new skills for user
-      tags: []
-      responses:
-        '201':
-          description: Created
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  statusCode:
-                    type: number
-                  body:
-                    $ref: ../models/User-Skill.yaml
-              examples:
-                Success:
-                  value:
-                    statusCode: 201
-                    body:
-                      id: asdfd
-                      skills:
-                        color-light:
-                          level: 1
-                          experiencePoints: 0
-                        anatomy:
-                          level: 1
-                          experiencePoints: 0
-                        composition:
-                          level: 1
-                          experiencePoints: 0
-                        construction:
-                          level: 1
-                          experiencePoints: 0
-                        design:
-                          level: 1
-                          experiencePoints: 0
-                        gesture:
-                          level: 1
-                          experiencePoints: 0
-                        perspective:
-                          level: 1
-                          experiencePoints: 0
-        '400':
-          description: Invalid request
-        '401':
-          description: Unauthorized
-        '500':
-          description: Internal Server Error
-      operationId: post-skills
-    patch:
-      summary: Update user skills by userid
-      tags: []
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: ../models/User-Skill.yaml
-              examples:
-                Success:
-                  value:
-                    id: asdfd
-                    skills:
-                      color-light:
-                        level: 1
-                        experiencePoints: 0
-                      anatomy:
-                        level: 1
-                        experiencePoints: 0
-                      composition:
-                        level: 1
-                        experiencePoints: 0
-                      construction:
-                        level: 1
-                        experiencePoints: 0
-                      design:
-                        level: 1
-                        experiencePoints: 0
-                      gesture:
-                        level: 1
-                        experiencePoints: 0
-                      perspective:
-                        level: 1
-                        experiencePoints: 0
-        '400':
-          description: Invalid request
-        '401':
-          description: Unauthorized
-        '500':
-          description: Internal Server Error
-      operationId: patch-skills
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                tags:
-                  type: array
-                  items:
-                    type: string
-            examples:
-              Update skills:
-                value:
-                  tags:
-                    - perspective
-                    - gesture
-          application/javascript:
-            schema:
-              type: object
-              properties:
-                tags:
-                  type: array
-                  items:
-                    type: string
-            examples:
-              Update skills:
-                value:
-                  tags:
-                    - perspective
-                    - gesture
-        description: ''
-    delete:
-      summary: Delete user skills by userid
-      tags: []
-      responses:
-        '200':
-          description: OK
-        '400':
-          description: Invalid request
-        '401':
-          description: Unauthorized
-        '500':
-          description: Internal Server Error
-      operationId: delete-skills
-    get:
-      summary: Get user skills by userid
-      tags: []
-      responses:
-        '200':
-          description: ''
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  statusCode:
-                    type: number
-                  body:
-                    $ref: ../models/User-Skill.yaml
-              examples:
-                Success:
-                  value:
-                    statusCode: 200
-                    body:
-                      id: asdfd
-                      skills:
-                        color-light:
-                          level: 1
-                          experiencePoints: 0
-                        anatomy:
-                          level: 1
-                          experiencePoints: 0
-                        composition:
-                          level: 1
-                          experiencePoints: 0
-                        construction:
-                          level: 1
-                          experiencePoints: 0
-                        design:
-                          level: 1
-                          experiencePoints: 0
-                        gesture:
-                          level: 1
-                          experiencePoints: 0
-                        perspective:
-                          level: 1
-                          experiencePoints: 0
-        '400':
-          description: Invalid request
-        '401':
-          description: Unauthorized
-        '500':
-          description: Internal Server Error
-      operationId: get-skills
-    parameters:
-      - schema:
-          type: string
-        name: userId
-        in: path
-        required: true
-components:
-  schemas: {}
