--- apps/skills-microservice/template.env
@@ -1,1 +1,0 @@
-IS_OFFLINE=true
