--- apps/skills-microservice/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["ESNEXT"],
-    "module": "CommonJS",
-    "moduleResolution": "node",
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "outDir": "build",
-    "removeComments": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "target": "ES5"
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": [
-    ".build/**/*",
-    ".dynamodb/**/*",
-    "node_modules/**/*",
-    ".serverless/**/*",
-    "_warmup/**/*",
-    ".vscode/**/*"
-  ]
-}
