--- apps/web/components/layout.tsx
@@ -1,33 +1,0 @@
-import Head from 'next/head'
-import Header from './header'
-
-export default function Layout({ user, loading = false, children }) {
-  return (
-    <>
-      <Head>
-        <title>Draw-App</title>
-      </Head>
-
-      <Header user={user} loading={loading} />
-
-      <main>
-        <div className="container">{children}</div>
-      </main>
-
-      <style jsx>{`
-        .container {
-          max-width: 42rem;
-          margin: 1.5rem auto;
-        }
-      `}</style>
-      <style jsx global>{`
-        body {
-          margin: 0;
-          color: #333;
-          font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-            Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-        }
-      `}</style>
-    </>
-  )
-}
