--- apps/web/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './auth0'
-export * from './user'
