+++ apps/web/package.json
@@ -6,10 +6,13 @@
     "dev": "npx next dev",
     "build": "npx next build",
     "start": "npx next start",
-    "lint": "npx next lint"
+    "lint": "npx next lint",
+    "test": "jest -c jest.config.ts --runInBand"
   },
   "dependencies": {
-    "@auth0/nextjs-auth0": "^1.7.0",
+    "@auth0/nextjs-auth0": "^1.8.0",
+    "aws-sdk": "^2.1058.0",
+    "axios": "^0.27.2",
     "next": "^12.0.10",
     "react": "17.0.2",
     "react-dom": "17.0.2",
