--- apps/web/pages/api/auth/logout.ts
@@ -1,10 +1,0 @@
-import auth0 from '../../../lib/auth0'
-
-export default async function logout(req, res) {
-  try {
-    await auth0.handleLogout(req, res)
-  } catch (error) {
-    console.error(error)
-    res.status(error.status || 500).end(error.message)
-  }
-}
