--- apps/web/pages/index.tsx
@@ -1,26 +1,0 @@
-import Layout from '../components/layout'
-import { useFetchUser } from '../lib'
-
-export default function Home() {
-  const { user, loading } = useFetchUser()
-
-  return (
-    <Layout user={user} loading={loading}>
-      <h1>Next.js and Auth0 Example</h1>
-
-      {loading && <p>Loading login info...</p>}
-
-      {!loading && !user && (
-        <>
-          <p>
-            To test the login click in <i>Login</i>
-          </p>
-          <p>
-            Once you have logged in you should be able to click in{' '}
-            <i>Profile</i> and <i>Logout</i>
-          </p>
-        </>
-      )}
-    </Layout>
-  )
-}
