+++ apps/web/src/components/profileCard.tsx
@@ -1,0 +1,7 @@
+export const ProfileCard = ({ user }) => (
+  <div>
+    <img src={user?.picture} alt="user picture" />
+    <p>nickname: {user?.nickname}</p>
+    <p>name: {user?.name}</p>
+  </div>
+)
