+++ apps/web/src/pages/_app.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { UserProvider } from '@auth0/nextjs-auth0'
+
+export default function App({ Component, pageProps }) {
+  return (
+    <UserProvider>
+      <Component {...pageProps} />
+    </UserProvider>
+  )
+}
