+++ apps/web/template.env
@@ -1,9 +1,16 @@
-SESSION_COOKIE_SECRET=
-NEXT_PUBLIC_AUTH0_DOMAIN=
-NEXT_PUBLIC_BASE_URL=
-NEXT_PUBLIC_AUTH0_CLIENT_ID=
-AUTH0_CLIENT_SECRET=
-NEXT_PUBLIC_REDIRECT_URI=
-NEXT_PUBLIC_POST_LOGOUT_REDIRECT_URI=
-NEXT_PUBLIC_AUTH0_SCOPE=
-SESSION_COOKIE_LIFETIME=
+# A long, secret value used to encrypt the session cookie
+AUTH0_SECRET=''
+
+# The base url of your application
+AUTH0_BASE_URL=''
+
+# The url of your Auth0 tenant domain
+AUTH0_ISSUER_BASE_URL=''
+
+# Your Auth0 application's Client ID
+AUTH0_CLIENT_ID=''
+
+# Your Auth0 application's Client Secret
+AUTH0_CLIENT_SECRET=''
+
+SKILLS_TABLE=''
